requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
annexremote==1.6.0
    # via datalad
attrs==23.1.0
    # via jsonschema
boto==2.49.0
    # via datalad
certifi==2023.7.22
    # via
    #   flywheel-sdk
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pygit2
chardet==5.2.0
    # via datalad
charset-normalizer==3.3.0
    # via requests
cryptography==41.0.4
    # via secretstorage
datalad==0.19.3
    # via -r requirements.in
distro==1.8.0
    # via datalad
dotty-dict==1.3.1
    # via flywheel-gear-toolkit
fasteners==0.19
    # via datalad
flywheel-bids==1.2.6
    # via -r requirements.in
flywheel-gear-toolkit==0.6.13
    # via
    #   -r requirements.in
    #   flywheel-bids
flywheel-gears==0.3.0
    # via flywheel-gear-toolkit
flywheel-sdk==16.19.2
    # via
    #   -r requirements.in
    #   flywheel-bids
future==0.18.3
    # via flywheel-bids
humanize==4.8.0
    # via datalad
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via keyring
importlib-resources==5.13.0
    # via flywheel-bids
iso8601==2.1.0
    # via datalad
jaraco-classes==3.3.0
    # via
    #   keyring
    #   keyrings-alt
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jsonschema==3.2.0
    # via
    #   flywheel-bids
    #   flywheel-gear-toolkit
    #   flywheel-gears
keyring==24.2.0
    # via datalad
keyrings-alt==5.0.0
    # via datalad
looseversion==1.3.0
    # via datalad
more-itertools==10.1.0
    # via jaraco-classes
msgpack==1.0.7
    # via datalad
packaging==23.2
    # via
    #   datalad
    #   flywheel-sdk
patool==1.12
    # via datalad
platformdirs==3.11.0
    # via datalad
pycparser==2.21
    # via cffi
pygit2==1.13.2
    # via -r requirements.in
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via flywheel-sdk
python-gitlab==3.15.0
    # via datalad
requests==2.31.0
    # via
    #   datalad
    #   flywheel-sdk
    #   python-gitlab
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via
    #   flywheel-sdk
    #   python-gitlab
rfc3987==1.3.8
    # via flywheel-gears
rtstatlib==1.0.0
    # via flywheel-bids
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   flywheel-sdk
    #   jsonschema
    #   python-dateutil
tqdm==4.66.1
    # via datalad
typing-extensions==4.8.0
    # via datalad
urllib3==2.0.7
    # via
    #   flywheel-bids
    #   flywheel-sdk
    #   requests
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
